~ fix 35 issues from third code analysis sweep
Security
~ canInteractWithBankSystem return value check in block GUIs
~ item duplication race in upload/download blocks
~ TOCTOU withdraw in terminal → lock-withdraw pattern
~ SetAccountIconAsync blocked from untrusted slaves

Performance
~ BankAccountChangeStream listener leak (stable callback ref)
~ ByteBuf leak in async forwarding decode

API
~ ItemID.getStack() @Nullable → @NotNull
+ BankUser.fromJson() + permissionValue serialization
~ BankPermission.hasAnyPermission != 0
~ IClientBank class → interface

CLI
~ bank_disableNotifications wrong message
~ money_send_user unit mismatch
~ null-safe getUserByUUID/getUserByName

UI/UX
~ BankUploadScreen ownership toggle
~ ATMScreen tick listener single-registration
~ AmountButtonGroup maxWidth Math.max

Internal
~ withdrawAndFillInventory incomplete futures
~ null player guards in async callbacks
~ ServerBank rollback flags + negative parsing
~ long→int overflow in DropItemsInPlayerInventoryRequest
~ null server guard in ServerInfoRequest.createInfo
~ ServerBankAccount.addUser dirty flag on update
~ removeEmptyBanks isItemIDNotRemovable check
~ ItemIDManager semantic ItemStack comparison
~ null level guards in block entities
~ MoneyItem.hashCode full long worth
~ SyncOpenGUIPacket constructor private

+++ common/src/main/java/net/kroia/banksystem/minecraft/item/custom/money/MoneyItem.java
@@ -65,7 +65,7 @@ public class MoneyItem extends Item{
 
     @Override
     public int hashCode() {
-        long worth = (int) worth();
+        long worth = worth();
         return (int) (worth ^ (worth >>> 32)) ^ getClass().hashCode(); // Use worth as hash code
     }
